refac class layout
