+++ Bugsnag/Assets/Bugsnag/Editor/SymbolUpload/BugsnagCLIDownloader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b2d8e93214749444c92dc4c184cd1533
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
